mcp_servers/python/codegen/pyproject.toml
[project]
name = "codegen-mcp-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.6.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]
